Made function and property styling more consistent.
Every class member has anchor links now. Added anchor links for
functions.

Made property and function names act as the anchor links, similar to
classes. As a consequence member names are blue now. This looks good to
me and is consistent with class names not hiding the fact that they're
anchor links, but that's subjective.