New CSS (#11)
* Update miniwiki CSS with changes from @KerriFair

Make the default miniwiki less dreary. A little more of a modern default
look.

- Smaller headers
- No gray boxes around classes
- Content source indicator on the left

The miniwiki seems to be the most public face dumpling has right now. It
is the most visual artifact of the data that it's really all about. So
let's dress it up!

* Made function and property styling more consistent.

Every class member has anchor links now. Added anchor links for
functions.

Made property and function names act as the anchor links, similar to
classes. As a consequence member names are blue now. This looks good to
me and is consistent with class names not hiding the fact that they're
anchor links, but that's subjective.

* Add anchor links to events and callbacks